+++ geode-core/src/integrationTest/java/org/apache/geode/TXJUnitTest.java
@@ -247,9 +247,9 @@ public class TXJUnitTest {
       cmtre.setUserAttribute("uaValue1");
       assertEquals("uaValue1", cmtre.getUserAttribute());
 
-      long txRollbackChanges = stats.getTxRollbackChanges();
-      long txCommitChanges = stats.getTxCommitChanges();
-      long txFailureChanges = stats.getTxFailureChanges();
+      int txRollbackChanges = stats.getTxRollbackChanges();
+      int txCommitChanges = stats.getTxCommitChanges();
+      int txFailureChanges = stats.getTxFailureChanges();
       this.txMgr.begin();
       Region.Entry txre = this.region.getEntry("uaKey");
       assertEquals(this.region, txre.getRegion());
@@ -299,9 +299,9 @@ public class TXJUnitTest {
     }
 
     {
-      long txRollbackChanges = stats.getTxRollbackChanges();
-      long txCommitChanges = stats.getTxCommitChanges();
-      long txFailureChanges = stats.getTxFailureChanges();
+      int txRollbackChanges = stats.getTxRollbackChanges();
+      int txCommitChanges = stats.getTxCommitChanges();
+      int txFailureChanges = stats.getTxFailureChanges();
       this.region.create("key1", "value1");
       this.txMgr.begin();
       this.region.invalidate("key1");
@@ -432,7 +432,7 @@ public class TXJUnitTest {
   @Test
   public void testTwoRegionTxs() throws CacheException {
     final CachePerfStats stats = this.cache.getCachePerfStats();
-    long txCommitChanges;
+    int txCommitChanges;
     TransactionId myTxId;
 
     AttributesFactory<String, String> attributesFactory = new AttributesFactory<>();
@@ -2739,7 +2739,7 @@ public class TXJUnitTest {
   }
 
   private void doNonTxInvalidateRegionOp(CachePerfStats stats) throws Exception {
-    long txRollbackChanges = stats.getTxRollbackChanges();
+    int txRollbackChanges = stats.getTxRollbackChanges();
     this.region.create("key1", "value1");
     this.region.create("key2", "value2");
     this.txMgr.begin();
@@ -2771,7 +2771,7 @@ public class TXJUnitTest {
   }
 
   private void doNonTxDestroyRegionOp(CachePerfStats stats) throws Exception {
-    long txRollbackChanges = stats.getTxRollbackChanges();
+    int txRollbackChanges = stats.getTxRollbackChanges();
     this.region.put("key1", "value1");
     this.region.put("key2", "value2");
     this.txMgr.begin();
@@ -4627,15 +4627,15 @@ public class TXJUnitTest {
       private long txSuccessLifeTime;
       private long txFailedLifeTime;
       private long txRollbackLifeTime;
-      private long txCommits;
-      private long txFailures;
-      private long txRollbacks;
+      private int txCommits;
+      private int txFailures;
+      private int txRollbacks;
       private long txCommitTime;
       private long txFailureTime;
       private long txRollbackTime;
-      private long txCommitChanges;
-      private long txFailureChanges;
-      private long txRollbackChanges;
+      private int txCommitChanges;
+      private int txFailureChanges;
+      private int txRollbackChanges;
 
       private CachePerfStats stats;
 
@@ -4670,15 +4670,15 @@ public class TXJUnitTest {
         this.txRollbackLifeTime = txRollbackLifeTime;
       }
 
-      private void setTxCommits(long txCommits) {
+      private void setTxCommits(int txCommits) {
         this.txCommits = txCommits;
       }
 
-      private void setTxFailures(long txFailures) {
+      private void setTxFailures(int txFailures) {
         this.txFailures = txFailures;
       }
 
-      private void setTxRollbacks(long txRollbacks) {
+      private void setTxRollbacks(int txRollbacks) {
         this.txRollbacks = txRollbacks;
       }
 
@@ -4694,15 +4694,15 @@ public class TXJUnitTest {
         this.txRollbackTime = txRollbackTime;
       }
 
-      private void setTxCommitChanges(long txCommitChanges) {
+      private void setTxCommitChanges(int txCommitChanges) {
         this.txCommitChanges = txCommitChanges;
       }
 
-      private void setTxFailureChanges(long txFailureChanges) {
+      private void setTxFailureChanges(int txFailureChanges) {
         this.txFailureChanges = txFailureChanges;
       }
 
-      private void setTxRollbackChanges(long txRollbackChanges) {
+      private void setTxRollbackChanges(int txRollbackChanges) {
         this.txRollbackChanges = txRollbackChanges;
       }
 
@@ -4831,10 +4831,10 @@ public class TXJUnitTest {
   public void testCacheStats() throws CacheException {
     CachePerfStats cacheStats = this.cache.getCachePerfStats();
     // quick sanity check to make sure perf stats work non-tx
-    long creates;
-    long destroys;
-    long puts;
-    long invalidates;
+    int creates;
+    int destroys;
+    int puts;
+    int invalidates;
 
     creates = cacheStats.getCreates();
     destroys = cacheStats.getDestroys();
@@ -5778,7 +5778,7 @@ public class TXJUnitTest {
 
     { // distributed invalidate
       // first make sure invalidate is counted as a change
-      long txRollbackChanges = stats.getTxRollbackChanges();
+      int txRollbackChanges = stats.getTxRollbackChanges();
       this.region.create("key1", "value1");
       this.txMgr.begin();
       this.region.invalidate("key1");
@@ -5837,7 +5837,7 @@ public class TXJUnitTest {
 
     { // local invalidate
       // first make sure invalidate is counted as a change
-      long txRollbackChanges = stats.getTxRollbackChanges();
+      int txRollbackChanges = stats.getTxRollbackChanges();
       this.region.create("key1", "value1");
       this.txMgr.begin();
       this.region.localInvalidate("key1");
